Add stuff to make tests.
Add a scripts key, and test command (run tests with 'npm test').

Initial test and mocks.

Lock versions to what we know we can/want to support, updates to Browserify's API and the produced module.

Rename to be more explicit.

Add a way to test the module by compiling it inline, but it depends on document.

Remove inline compling for now.